requirements.txt
asgiref
Django
gunicorn
packaging
Pillow
psycopg2-binary
python-decouple
sqlparse
tzdata
whitenoise
